analytics: move 'watchman integration (6799)' into sub-folder
Author: mayeaux <contact@anthonymayfield.com>
Mover: infinite-persistence <inf.persistence@gmail.com>

Moved items as-is, with the exception of adding `gWatchmanAnalyticsEnabled` instead of hardcoding to `IS_WEB`.
