+++ ui/analytics.js
@@ -2,7 +2,8 @@
 import { Lbryio } from 'lbryinc';
 import * as Sentry from '@sentry/browser';
 import * as RENDER_MODES from 'constants/file_render_modes';
-import { SDK_API_PATH } from 'config';
+import { watchman } from 'analytics/watchman';
+import type { Watchman } from 'analytics/watchman';
 
 // --- GA ---
 // - Events: 500 max (cannot be deleted).
@@ -33,27 +34,15 @@ export const GA_DIMENSIONS = {
   END_TIME_MS: 'end_time_ms',
 };
 
-// import getConnectionSpeed from 'util/detect-user-bandwidth';
-
-// let userDownloadBandwidthInBitsPerSecond;
-// async function getUserBandwidth() {
-//   userDownloadBandwidthInBitsPerSecond = await getConnectionSpeed();
-// }
-
-// get user bandwidth every minute, starting after an initial one minute wait
-// setInterval(getUserBandwidth, 1000 * 60);
-
 const isProduction = process.env.NODE_ENV === 'production';
 let gAnalyticsEnabled = false;
 const devInternalApis = process.env.LBRY_API_URL && process.env.LBRY_API_URL.includes('dev');
-
-const WATCHMAN_BACKEND_ENDPOINT = 'https://watchman.na-backend.odysee.com/reports/playback';
-const SEND_DATA_TO_WATCHMAN_INTERVAL = 10; // in seconds
 
 type Analytics = {
   setState: (enable: boolean) => void,
   appStartTime: number,
   eventStartTime: any,
+  video: Watchman,
   error: (string) => Promise<any>,
   sentryError: ({} | string, {}) => Promise<any>,
   setUser: (Object) => void,
@@ -62,21 +51,6 @@ type Analytics = {
   tagFollowEvent: (string, boolean, ?string) => void,
   playerLoadedEvent: (string, ?boolean) => void,
   playerVideoStartedEvent: (?boolean) => void,
-  videoStartEvent: (?string, number, string, ?number, string, any, ?number, boolean) => void,
-  videoIsPlaying: (boolean, any) => void,
-  videoBufferEvent: (
-    StreamClaim,
-    {
-      timeAtBuffer: number,
-      bufferDuration: number,
-      bitRate: number,
-      duration: number,
-      userId: string,
-      playerPoweredBy: string,
-      readyState: number,
-      isLivestream: boolean,
-    }
-  ) => Promise<any>,
   adsFetchedEvent: () => void,
   emailProvidedEvent: () => void,
   emailVerifiedEvent: () => void,
@@ -99,202 +73,14 @@ type LogPublishParams = {
 
 const isGaAllowed = false; // internalAnalyticsEnabled && isProduction; -- TODO: will be moved in upcoming commit
 
-/**
- * Determine the mobile device type viewing the data
- * This function returns one of 'and' (Android), 'ios', or 'web'.
- *
- * @returns {String}
- */
-function getDeviceType() {
-  // We may not care what the device is if it's in a web browser. Commenting out for now.
-  // if (!IS_WEB) {
-  //   return 'elt';
-  // }
-  // const userAgent = navigator.userAgent || navigator.vendor || window.opera;
-  //
-  // if (/android/i.test(userAgent)) {
-  //   return 'adr';
-  // }
-  //
-  // // iOS detection from: http://stackoverflow.com/a/9039885/177710
-  // if (/iPad|iPhone|iPod/.test(userAgent) && !window.MSStream) {
-  //   return 'ios';
-  // }
-
-  // default as web, this can be optimized
-  if (!IS_WEB) {
-    return 'dsk';
-  }
-  return 'web';
-}
-// variables initialized for watchman
-let amountOfBufferEvents = 0;
-let amountOfBufferTimeInMS = 0;
-let videoType, userId, claimUrl, playerPoweredBy, videoPlayer, bitrateAsBitsPerSecond, isLivestream;
-let lastSentTime;
-
-// calculate data for backend, send them, and reset buffer data for next interval
-async function sendAndResetWatchmanData() {
-  if (!userId) {
-    return 'Can only be used with a user id';
-  }
-
-  if (!videoPlayer) {
-    return 'Video player not initialized';
-  }
-
-  let timeSinceLastIntervalSend = new Date() - lastSentTime;
-  lastSentTime = new Date();
-
-  let protocol;
-  if (videoType === 'application/x-mpegURL' && !isLivestream) {
-    protocol = 'hls';
-    // get bandwidth if it exists from the texttrack (so it's accurate if user changes quality)
-    // $FlowFixMe
-    bitrateAsBitsPerSecond = videoPlayer.tech(true).vhs?.playlists?.media?.().attributes?.BANDWIDTH;
-  } else if (isLivestream) {
-    protocol = 'lvs';
-    // $FlowFixMe
-    bitrateAsBitsPerSecond = videoPlayer.tech(true).vhs?.playlists?.media?.().attributes?.BANDWIDTH;
-  } else {
-    protocol = 'stb';
-  }
-
-  // current position in video in MS
-  const positionInVideo = isLivestream ? 0 : videoPlayer && Math.round(videoPlayer.currentTime()) * 1000;
-
-  // get the duration marking the time in the video for relative position calculation
-  const totalDurationInSeconds = isLivestream ? 0 : videoPlayer && Math.round(videoPlayer.duration());
-
-  // temp: if buffering over the interval, the duration doesn't reset since we don't get an event
-  if (amountOfBufferTimeInMS > timeSinceLastIntervalSend) amountOfBufferTimeInMS = timeSinceLastIntervalSend;
-
-  // build object for watchman backend
-  const objectToSend = {
-    rebuf_count: amountOfBufferEvents,
-    rebuf_duration: amountOfBufferTimeInMS,
-    url: claimUrl.replace('lbry://', ''),
-    device: getDeviceType(),
-    duration: timeSinceLastIntervalSend,
-    protocol,
-    player: playerPoweredBy,
-    user_id: userId.toString(),
-    position: isLivestream ? 0 : Math.round(positionInVideo),
-    rel_position: isLivestream ? 0 : Math.round((positionInVideo / (totalDurationInSeconds * 1000)) * 100),
-    bitrate: bitrateAsBitsPerSecond,
-    bandwidth: undefined,
-    // ...(userDownloadBandwidthInBitsPerSecond && {bandwidth: userDownloadBandwidthInBitsPerSecond}), // add bandwidth if populated
-  };
-
-  // post to watchman
-  await sendWatchmanData(objectToSend);
-
-  // reset buffer data
-  amountOfBufferEvents = 0;
-  amountOfBufferTimeInMS = 0;
-}
-
-let watchmanInterval;
-// clear watchman interval and mark it as null (when video paused)
-function stopWatchmanInterval() {
-  clearInterval(watchmanInterval);
-  watchmanInterval = null;
-}
-
-// creates the setInterval that will run send to watchman on recurring basis
-function startWatchmanIntervalIfNotRunning() {
-  if (!watchmanInterval) {
-    // instantiate the first time to calculate duration from
-    lastSentTime = new Date();
-
-    // only set an interval if analytics are enabled and is prod
-    if (isProduction && IS_WEB) {
-      watchmanInterval = setInterval(sendAndResetWatchmanData, 1000 * SEND_DATA_TO_WATCHMAN_INTERVAL);
-    }
-  }
-}
-
-// post data to the backend
-async function sendWatchmanData(body) {
-  try {
-    const response = await fetch(WATCHMAN_BACKEND_ENDPOINT, {
-      method: 'POST',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(body),
-    });
-    return response;
-  } catch (err) {}
-}
-
 const analytics: Analytics = {
   setState: (enable: boolean) => {
     gAnalyticsEnabled = enable;
+    analytics.video.setState(gAnalyticsEnabled);
   },
   appStartTime: 0,
   eventStartTime: {},
-
-  // receive buffer events from tracking plugin and save buffer amounts and times for backend call
-  videoBufferEvent: async (claim, data) => {
-    amountOfBufferEvents = amountOfBufferEvents + 1;
-    amountOfBufferTimeInMS = amountOfBufferTimeInMS + data.bufferDuration;
-  },
-  onDispose: () => {
-    stopWatchmanInterval();
-  },
-  /**
-   * Is told whether video is being started or paused, and adjusts interval accordingly
-   * @param {boolean} isPlaying - Whether video was started or paused
-   * @param {object} passedPlayer - VideoJS Player object
-   */
-  videoIsPlaying: (isPlaying, passedPlayer) => {
-    let playerIsSeeking = false;
-    // have to use this because videojs pauses/unpauses during seek
-    // sometimes the seeking function isn't populated yet so check for it as well
-    if (passedPlayer && passedPlayer.seeking) {
-      playerIsSeeking = passedPlayer.seeking();
-    }
-
-    // if being paused, and not seeking, send existing data and stop interval
-    if (!isPlaying && !playerIsSeeking) {
-      sendAndResetWatchmanData();
-      stopWatchmanInterval();
-      // if being told to pause, and seeking, send and restart interval
-    } else if (!isPlaying && playerIsSeeking) {
-      sendAndResetWatchmanData();
-      stopWatchmanInterval();
-      startWatchmanIntervalIfNotRunning();
-      // is being told to play, and seeking, don't do anything,
-      // assume it's been started already from pause
-    } else if (isPlaying && playerIsSeeking) {
-      // start but not a seek, assuming a start from paused content
-    } else if (isPlaying && !playerIsSeeking) {
-      startWatchmanIntervalIfNotRunning();
-    }
-  },
-  videoStartEvent: (
-    claimId,
-    timeToStartVideo,
-    poweredBy,
-    passedUserId,
-    canonicalUrl,
-    passedPlayer,
-    videoBitrate,
-    isLivestreamClaim
-  ) => {
-    // populate values for watchman when video starts
-    userId = passedUserId;
-    claimUrl = canonicalUrl;
-    playerPoweredBy = poweredBy;
-    isLivestream = isLivestreamClaim;
-
-    videoType = passedPlayer.currentSource().type;
-    videoPlayer = passedPlayer;
-    bitrateAsBitsPerSecond = videoBitrate;
-    !isLivestreamClaim && sendPromMetric('time_to_start', timeToStartVideo, playerPoweredBy);
-  },
+  video: watchman,
   error: (message) => {
     return new Promise((resolve) => {
       if (gAnalyticsEnabled && isProduction) {
@@ -479,15 +265,6 @@ const analytics: Analytics = {
 function sendGaEvent(event: string, params?: { [string]: string | number }) {
   if (isGaAllowed && window.gtag) {
     window.gtag('event', event, params);
-  }
-}
-
-function sendPromMetric(name: string, value?: number, player: string) {
-  if (IS_WEB) {
-    let url = new URL(SDK_API_PATH + '/metric/ui');
-    const params = { name: name, value: value ? value.toString() : '', player: player };
-    url.search = new URLSearchParams(params).toString();
-    return fetch(url, { method: 'post' }).catch(function (error) {});
   }
 }
